+++ src/order/order_client.hpp
@@ -19,7 +19,7 @@ namespace finam::order {
 
 struct OrderState {
     std::string order_id;
-    int32_t     local_id{};
+    int64_t     local_id{};   // FIX: int64_t (Finam order IDs are 64-bit)
     Symbol      symbol;
     OrderSide   side{};
     OrderStatus status{OrderStatus::Pending};
@@ -35,9 +35,6 @@ using OrderUpdateCallback = std::function<void(const OrderUpdate&)>;
 
 class OrderClient : public IOrderExecutor {
 public:
-    // Конструктор без callback — его можно установить позже через set_update_callback().
-    // Типичный use-case: main создаёт OrderClient, затем StrategyRunner,
-    // затем регистрирует callback который знает про runner.
     explicit OrderClient(
         std::shared_ptr<auth::TokenManager> token_mgr,
         std::string                         account_id,
@@ -48,21 +45,17 @@ public:
     OrderClient(const OrderClient&)            = delete;
     OrderClient& operator=(const OrderClient&) = delete;
 
-    // ── IOrderExecutor ────────────────────────────────────────────────────────────────
-    [[nodiscard]] Result<int32_t> submit(const OrderRequest& req) override;
+    // ── IOrderExecutor ────────────────────────────────────────────────────────
+    // FIX: Result<int64_t> — соответствует IOrderExecutor и Finam API
+    [[nodiscard]] Result<int64_t> submit(const OrderRequest& req) override;
     [[nodiscard]] Result<void>    cancel(int64_t order_no,
                                          std::string_view client_id) override;
 
-    // ── Query ─────────────────────────────────────────────────────────────────────
+    // ── Query ─────────────────────────────────────────────────────────────────
     [[nodiscard]] std::vector<OrderState>   active_orders() const;
-    [[nodiscard]] std::optional<OrderState> find(int32_t local_id) const;
+    [[nodiscard]] std::optional<OrderState> find(int64_t local_id) const; // FIX: int64_t
 
-    // ── Callback ─────────────────────────────────────────────────────────────────
-    //
-    // Заменяет callback атомарно — безопасно вызывать до первого
-    // прихода данных из order stream. Stream стартует в конструкторе,
-    // но первый ответ от биржи приходит через несколько мс — окно
-    // в main достаточно узкое.
+    // ── Callback ──────────────────────────────────────────────────────────────
     void set_update_callback(OrderUpdateCallback cb) noexcept {
         std::lock_guard lock(cb_mu_);
         on_update_ = std::move(cb);
@@ -72,13 +65,11 @@ public:
 
 private:
     [[nodiscard]] std::unique_ptr<grpc::ClientContext> make_context() const;
-    [[nodiscard]] int32_t next_id() noexcept;
+    [[nodiscard]] int64_t next_id() noexcept; // FIX: int64_t
 
     void upsert(OrderState state);
     void run_order_stream();
 
-    // Инвокация on_update_ через обёртку — защищает замену каллбэка
-    // посреди его вызова.
     void invoke_callback(const OrderUpdate& upd) {
         std::lock_guard lock(cb_mu_);
         if (on_update_) on_update_(upd);
@@ -87,14 +78,14 @@ private:
     std::shared_ptr<auth::TokenManager> token_mgr_;
     std::string                         account_id_;
 
-    mutable std::mutex  cb_mu_;       // защищает on_update_
-    OrderUpdateCallback on_update_;   // защищён cb_mu_
+    mutable std::mutex  cb_mu_;
+    OrderUpdateCallback on_update_;
 
     std::unique_ptr<::grpc::tradeapi::v1::orders::OrdersService::Stub> orders_stub_;
 
-    std::atomic<int32_t>                    id_counter_{1};
+    std::atomic<int64_t>                    id_counter_{1};  // FIX: int64_t
     mutable std::shared_mutex               orders_mu_;
-    std::unordered_map<int32_t, OrderState> orders_;
+    std::unordered_map<int64_t, OrderState> orders_;         // FIX: int64_t key
 
     std::atomic<bool> stop_{false};
     std::thread       stream_thread_;
